Improving functions interoperability
